refactor(cpuhealth): table_cpu.go holds the CPU table builder
table.go now keeps only the signal/track name constants, the tiers,
and the two entry points (NewEngine, Table). The function that
assembles them into a diagnosis.Table lives in its own file, named for
what it holds rather than mirrored off the per-signal file naming
(cpuTable declares the whole table, not one signal).
